Create ShowList_ResourcesGroup.sh
To export the details in CSV files using Azure CLI, you can use the --query option with the --output option to format the output as a table and then redirect it to a CSV file. Here's an example of how to do it for the "Retrieve a list of resource groups" scenario:

Retrieve a list of resource groups in your subscription and export to a CSV file:

In this example, we're using the --output table option to format the output as a table and the --query option to specify the fields we want to include in the CSV file. We're selecting the "name" and "location" fields and redirecting the output to a CSV file named "ResourceGroups.csv".

You can customize the --query parameter to include additional fields or format the output as needed.

Repeat a similar approach for other Azure CLI commands to export the output in CSV format.