Update Show_SubscripationInfo.sh
To export the details in CSV format instead of a JSON file using Azure CLI, you can follow a similar approach as in the previous example. Here's how to retrieve information about the current Azure subscription and export it to a CSV file:

In this example, we're using the --output table option to format the output as a table and the --query option to specify the fields we want to include in the CSV file. We're selecting the "name," "id" (Subscription ID), "tenantId," and "isDefault" fields and redirecting the output to a CSV file named "SubscriptionInfo.csv."

You can customize the --query parameter to include additional fields or format the output as needed. This approach allows you to export the information in CSV format instead of a JSON file.